per review: move action handler file to prim
