+++ stan/math/prim/mat/fun/matrix_exp_multiply.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_PRIM_MAT_FUN_MATRIX_EXP_MULTIPLY_HPP
 #define STAN_MATH_PRIM_MAT_FUN_MATRIX_EXP_MULTIPLY_HPP
 
-#include <stan/math/rev/mat/fun/matrix_exp_action_handler.hpp>
+#include <stan/math/prim/mat/fun/matrix_exp_action_handler.hpp>
 
 namespace stan {
 namespace math {
