+++ test/unit/math/rev/mat/fun/matrix_exp_action_handler_test.cpp
@@ -2,7 +2,7 @@
 // #include <stan/math/prim/scal/err/check_not_nan.hpp>
 #include <test/unit/math/prim/mat/util.hpp>
 #include <stan/math/prim/mat/fun/matrix_exp.hpp>
-#include <stan/math/rev/mat/fun/matrix_exp_action_handler.hpp>
+#include <stan/math/prim/mat/fun/matrix_exp_action_handler.hpp>
 #include <vector>
 
 TEST(MathMatrix, matrix_exp_action_diag) {
